requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0; python_version >= '3.7'
asgiref==3.8.1; python_version >= '3.8'
astunparse==1.6.3
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
dj-database-url==2.1.0
django==5.0.6; python_version >= '3.10'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
flatbuffers==24.3.25
gast==0.5.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.63.0; python_version >= '3.8'
gunicorn==22.0.0; python_version >= '3.7'
h5py==3.11.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
keras==3.3.3; python_version >= '3.9'
libclang==18.1.1
markdown==3.6; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
ml-dtypes==0.3.2; python_version >= '3.9'
mysqlclient==2.2.4; python_version >= '3.8'
namex==0.0.8
numpy==1.26.4; python_version <= '3.11'
opt-einsum==3.3.0; python_version >= '3.5'
optree==0.11.0; python_version >= '3.7'
packaging==24.0; python_version >= '3.7'
pillow==10.3.0; python_version >= '3.8'
protobuf==4.25.3; python_version >= '3.8'
# psycopg2==2.9.9; python_version >= '3.7'
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
rich==13.7.1; python_full_version >= '3.7.0'
setuptools==69.5.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.5.0; python_version >= '3.8'
tensorboard==2.16.2; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
tensorflow==2.16.1; python_version >= '3.9'
tensorflow-intel==2.16.1; platform_system == 'Windows'
tensorflow-io-gcs-filesystem==0.31.0; python_version < '3.12'
termcolor==2.4.0; python_version >= '3.8'
typing-extensions==4.11.0; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
urllib3==2.2.1; python_version >= '3.8'
werkzeug==3.0.3; python_version >= '3.8'
wheel==0.43.0; python_version >= '3.8'
wrapt==1.16.0; python_version >= '3.6'
